04 auth

+++ controllers/contactsControllers.js
@@ -9,7 +9,7 @@ import { isValidObjectId } from 'mongoose';
 
 export const getAllContacts = async (req, res) => {
     try {
-        const arrContacts = await Contact.find();
+        const arrContacts = await Contact.find({ owner: req.user.id });
         res.json(arrContacts);
     } catch (error) {
         next(error);
@@ -19,13 +19,17 @@ export const getAllContacts = async (req, res) => {
 export const getOneContact = async (req, res, next) => {
     try {
         const { id } = req.params;
+
         if (!isValidObjectId(id)) {
             throw HttpError(400, `${id} is not valid id`);
         }
-        const contactById = await Contact.findById(id);
-        if (!contactById) {
-            throw HttpError(404);
+
+        const contactById = await Contact.findOne({ _id: id, owner: req.user.id });
+
+        if (contactById === null) {
+            throw HttpError(401, 'Not authorized');
         }
+
         res.json(contactById);
     } catch (error) {
         next(error);
@@ -38,7 +42,7 @@ export const deleteContact = async (req, res, next) => {
         if (!isValidObjectId(id)) {
             throw HttpError(400, `${id} is not valid id`);
         }
-        const removeContact = await Contact.findByIdAndDelete(id);
+        const removeContact = await Contact.findByIdAndDelete({ _id: id, owner: req.user.id });
         if (!removeContact) {
             throw HttpError(404);
         }
@@ -55,7 +59,16 @@ export const createContact = async (req, res, next) => {
         if (error) {
             throw HttpError(400, error.message);
         }
-        const newAddContact = await Contact.create(req.body);
+
+        const contact = {
+            name: req.body.name,
+            email: req.body.email,
+            phone: req.body.phone,
+            owner: req.user.id,
+            favorite: req.body.favorite,
+        };
+        const newAddContact = await Contact.create(contact);
+
         res.status(201).json(newAddContact);
     } catch (error) {
         next(error);
@@ -76,7 +89,11 @@ export const updateContact = async (req, res, next) => {
             throw HttpError(400, `${id} is not valid id`);
         }
 
-        const newUpdateContact = await Contact.findByIdAndUpdate(id, req.body, { new: true });
+        const newUpdateContact = await Contact.findByIdAndUpdate(
+            { _id: id, owner: req.user.id },
+            req.body,
+            { new: true }
+        );
 
         if (!newUpdateContact) {
             throw HttpError(404);
@@ -102,7 +119,11 @@ export const updateStatusContact = async (req, res, next) => {
             throw HttpError(400, `${contactId} is not valid id`);
         }
 
-        const newUpdateStatus = await Contact.findByIdAndUpdate(contactId, req.body, { new: true });
+        const newUpdateStatus = await Contact.findByIdAndUpdate(
+            { _id: contactId, owner: req.user.id },
+            req.body,
+            { new: true }
+        );
 
         if (!newUpdateStatus) {
             throw HttpError(404);
